Set the stopping criteria in a form of the relative changes during fitting iterations.

Set the stopping criteria in a form of the relative changes during fitting iterations.

setTolerance(params = 0.01, shared = 0.01, normFactors = 0.01,
  options = .defaultParams)

Arguments

params
a threshold for gene-specific parameter boundaries

shared
a threshold for shared parameters boundaries

normFactors
a threshold for the fraction factors

options
an options object to use as a basis for a new parameter set

Value

an options object with the new parameter values

Details

If no options object is provided, the default options are used as a base. A threshold represents the relative changes in parameter values between two subsequent fitting iterations.

Examples

setTolerance(params = 1e-2)

#> $tolerance
#> $tolerance$params
#> [1] 0.01
#> 
#> $tolerance$shared
#> [1] 0.01
#> 
#> $tolerance$normFactors
#> [1] 0.001
#> 
#> 
#> $verbose
#> [1] "silent"
#> 
#> $lb
#> $lb$size
#> [1] 10
#> 
#> 
#> $ub
#> $ub$size
#> [1] 1e+10
#> 
#>