Create a likelihood function

Create a likelihood function

totalll(par, pd, include = names(par))

Arguments

par
a named list with parameter values (including the size parameter for the negative binomial distribution, see dnbinom)

pd
a PulseData object

include
character vector, names of the parameters in the par list, which will be used in the created likelihood function

Value

a function with the following arguments:

x, a numeric vector which correspods to the records in par argument. The first element of the normFactors, if par$normFactors is not NULL, is excluded, because it is assumed to be fixed to 1. Other parameters are the same as in par. The structure of x is identical to unlist(par[include]). The first element of par$normFactors is removed.

The created function returns a minus logarithm of the likelihood function calculated on the basis of the negative binomial distribution for the provided counts, normalisation factors and parameters.