Create an object for pulse-change count data

Create an object for pulse-change count data

PulseData(counts, conditions, formulas, formulaIndexes = NULL,
  spikeins = NULL, groups = NULL)

Arguments

counts
a matrix; column names correspond to sample names. The columns in counts correspond to the rows in conditions argument.

conditions
a data.frame; the first column corresponds to the conditions given in formulas. The order of rows corresponds to the columns (samples) in the counts argument.

formulas
a list, created by MeanFormulas

formulaIndexes
a list of lists; defines indexes of formulas used for calculation of the expected read number

spikeins
NULL (default) or a list of two items:

refGroup, a character, which defines the group which should be treated as a reference for normalisation

spikeLists, a list of character vectors with the spike-ins names and the same structure as formulaIndexes.

groups
NULL (default) or a formula or a vector, e.g. ~ condition + time (if spike-ins are not provided). The vector length must be the same as the sample number. the names used in the fractions defines different fractions, which should have distinct coefficients for mean expression fitting.

Value

an object of class "PulseData"

Details

The conditions argument may include additional columns, which provide values for known parameters, such as time. Their name must be the same as defined in formulas. For example, if a formula is defined as mu * exp(-d * time) where time is the time point of the experiment, the condition data.frame must contain a column named time.

Examples

# Spike-ins definition for object creation

formulaIndexes <- list(
  total_fraction = 'total',
  flow_through   = c('unlabelled', 'labelled'),
  pull_down      = c('labelled', 'unlabelled'))

# spike-ins set up:

refGroup <- "total_fraction"

labelled <- c("spike1", "spike2")
unlabelled <- c("spike3", "spike4")

spikeLists <- list(
#  total samples are normalised using all spike-ins
  total_fraction = list(c(unlabelled, labelled)),
# for every item in formulaIndexes we have a set of spike-ins:   
  flow_through   = list(unlabelled, labelled),
  pull_down      = list(labelled, unlabelled))

# argument for the function: 
spikeins <- list(refGroup = refGroup,
                 spikeLists = spikeLists)