Estimate profile likelihood for gene parameters (all other fixed)

Estimate profile likelihood for gene parameters (all other fixed)

profileGene(parName, geneIndex, pd, par, interval, options, ...)

profile(paramPath, pd, par, options, interval, namesToOptimise = names(par),
  ...)

Arguments

parName
a character; the names of the gene-specific parameter (e.g. "mu")

geneIndex
integer(1); the row index in the count table where the data for the given gene are located

pd
the PulseData object

par
result output from the fitModel function

interval
a vector of two numbers defining the interval of parameter profiling

options
the options list used for the fitModel function call

...
other parameters to pass to runPL, i.e. logScale (logical) and number of points numPoints

paramPath
a list with names and indexes in order to locate the parameter in the par argument (e.g. list("mu", 1) corresponds to the "mu" parameter value for the first gene, i.e. par[["mu"]][[1]].

namesToOptimise
a character vector including the free parameters (i.e. they wil be treated as unfixed)

Value

a data.frame; the column logl corresponds to the -log(likelihood) function values. the other represents the profiled parameter values.